refactor enums to separate source files
