+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/LoRaMessage/LoRaPayload.cs
@@ -12,33 +12,6 @@ namespace LoRaTools.LoRaMessage
     using Org.BouncyCastle.Crypto.Parameters;
     using Org.BouncyCastle.Security;
 
-    public enum LoRaPayloadKeyType
-    {
-        NwkSkey = 1,
-        AppSKey = 2
-    }
-
-    public enum LoRaMessageType : byte
-    {
-        // Request sent by device to join
-        JoinRequest,
-
-        // Response to a join request sent to device
-        JoinAccept = 32,
-
-        // Device to cloud message, no confirmation expected
-        UnconfirmedDataUp = 64,
-
-        // Cloud to device message, no confirmation expected
-        UnconfirmedDataDown = 96,
-
-        // Device to cloud message, confirmation required
-        ConfirmedDataUp = 128,
-
-        // Cloud to device message, confirmation required
-        ConfirmedDataDown = 160
-    }
-
     /// <summary>
     /// The LoRaPayloadWrapper class wraps all the information any LoRa message share in common
     /// </summary>
