+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/LoRaMessage/LoRaPayloadKeyType.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace LoRaTools.LoRaMessage
+{
+    public enum LoRaPayloadKeyType
+    {
+        NwkSkey = 1,
+        AppSKey = 2
+    }
+}
