+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/Mac/CidEnum.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace LoRaTools
+{
+    public enum CidEnum
+    {
+        Zero,
+        One,
+        LinkCheckCmd,
+        LinkADRCmd,
+        DutyCycleCmd,
+        RXParamCmd,
+        DevStatusCmd,
+        NewChannelCmd,
+        RXTimingCmd
+    }
+}
