+++ LoRaEngine/modules/LoRaWanNetworkSrvModule/LoraTools/Regions/LoRaRegion.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+namespace LoRaTools.Regions
+{
+    public enum LoRaRegion
+    {
+        EU868,
+        US915
+    }
+}
